docs: Go house style guide + lint hardening (#927) (#928)
Adds docs/go-style.md — the repo's Go house style: rationale-only
comments citing spec/tool/device, operator-sentence errors, the
honesty doctrine (no silent fallbacks; tests assert from the
authority), and the linting policy (a linter is only enabled when
the whole tree already passes it).

Enables misspell, unconvert, wastedassign in .golangci.yml and
makes the tree pass them: 19 redundant conversions removed and 4
wasted assignments dropped across cmd/dhs, acp2, amwa, cerebrum-nb,
emberplus, osc, probel-sw02p, export.

+++ internal/acp2/provider/encoder.go
@@ -444,7 +444,7 @@ func stringMaxLen(p *canonical.Parameter) uint16 {
 		return 256
 	}
 	if ml := maxLenHint(p); ml > 0 {
-		return uint16(ml)
+		return ml
 	}
 	if p.Maximum != nil {
 		if u, err := asUint32(p.Maximum, "max"); err == nil && u > 0 && u <= 0xFFFF {
